+++ core/network/src/main/java/org/cazait/network/api/AuthedService.kt
@@ -1,0 +1,17 @@
+package org.cazait.network.api
+
+import org.cazait.network.model.dto.request.CafeReviewPostReq
+import org.cazait.network.model.dto.response.CafeReviewPostRes
+import retrofit2.Response
+import retrofit2.http.Body
+import retrofit2.http.POST
+import retrofit2.http.Path
+
+interface AuthedService {
+    @POST("/api/reviews/user/{userId}/cafe/{cafeId}")
+    suspend fun postReviewAuth(
+        @Path("userId") userId: Long,
+        @Path("cafeId") cafeId: Long,
+        @Body reviewPostReq: CafeReviewPostReq,
+    ): Response<CafeReviewPostRes>
+}
