+++ core/network/src/main/java/org/cazait/network/api/CafeService.kt
@@ -80,12 +80,12 @@ interface CafeService {
         @Path("cafeId") cafeId: Long,
     ): Response<DeleteFavoriteCafeRes>
 
-    @POST("/api/reviews/user/{userId}/cafe/{cafeId}")
-    suspend fun postReviewAuth(
-        @Path("userId") userId: Long,
-        @Path("cafeId") cafeId: Long,
-        @Body reviewPostReq: CafeReviewPostReq,
-    ): Response<CafeReviewPostRes>
+//    @POST("/api/reviews/user/{userId}/cafe/{cafeId}")
+//    suspend fun postReviewAuth(
+//        @Path("userId") userId: Long,
+//        @Path("cafeId") cafeId: Long,
+//        @Body reviewPostReq: CafeReviewPostReq,
+//    ): Response<CafeReviewPostRes>
 
     @GET("/api/cafes/name/{cafeName}")
     suspend fun getCafeSearch(
