+++ core/network/src/main/java/org/cazait/network/datasource/CafeInfoRemoteData.kt
@@ -2,15 +2,12 @@ package org.cazait.network.datasource
 
 import org.cazait.network.NetworkConnectivity
 import org.cazait.network.api.CafeService
+import org.cazait.network.api.AuthedService
 import org.cazait.network.di.Authenticated
-import org.cazait.network.di.Unauthenticated
 import org.cazait.network.error.NETWORK_ERROR
-import org.cazait.network.error.NO_INTERNET_CONNECTION
-import org.cazait.network.model.dto.CafeDTO
 import org.cazait.network.model.dto.DataResponse
 import org.cazait.network.model.dto.request.CafeReviewPostReq
 import org.cazait.network.model.dto.response.CafeMenuRes
-import org.cazait.network.model.dto.response.CafeRes
 import org.cazait.network.model.dto.response.CafeResTemp
 import org.cazait.network.model.dto.response.CafeReviewPostRes
 import org.cazait.network.model.dto.response.CafeReviewRes
@@ -21,12 +18,12 @@ import java.io.IOException
 import javax.inject.Inject
 
 class CafeInfoRemoteData @Inject constructor(
-    @Unauthenticated private val cafeService: CafeService,
-    @Authenticated private val cafeServiceAuth: CafeService,
+    private val cafeService: CafeService,
+    @Authenticated private val cafeServiceAuth: AuthedService,
     private val networkConnectivity: NetworkConnectivity,
 ) : CafeInfoRemoteDataSource {
     override suspend fun getCafe(cafeId: Long): DataResponse<CafeResTemp> {
-        return when(val response = processCall {
+        return when (val response = processCall {
             cafeService.getCafe(cafeId)
         }) {
             is CafeResTemp -> {
@@ -82,13 +79,14 @@ class CafeInfoRemoteData @Inject constructor(
         score: Int,
         content: String
     ): DataResponse<CafeReviewPostRes> {
-        return when(val response = processCall {
+        return when (val response = processCall {
             val req = CafeReviewPostReq(score, content)
             cafeServiceAuth.postReviewAuth(userId, cafeId, req)
         }) {
             is CafeReviewPostRes -> {
                 DataResponse.Success(data = response)
             }
+
             else -> {
                 DataResponse.DataError(errorCode = response as Int)
             }
